Decrease dependency on Vulkan loader.
This largely removes any dependency on Vulkan in the use of vkfl
objects. The required types (3 function pointers, 2 opaque type
definitions) are provided by the header itself now. This eliminates a
need to worry about how vulkan.h is or is not included. Additionally,
macros controlling the inclusion or exclusion of symbols (especially
symbols related to Vulkan extensions) have been removed. To further
improve things the snake_case function_pointer namespace is removed and
vkfl::command enum values are now PascalCase. This makes it easier to
wrap loading with a macro such as the ones found in the test programs.

Also some cleanup related to changing "vulkan-function-loader" to vkfl
is included.