Improve generation script.
This improves on the generate.py script in a variety of ways. Several
passes over the command list are now removed as well as some other
extraneous code. Generated headers now contain macros for detecting the
list of extensions and Vulkan API versions requested at build time. Also
included are macros for the date and time (UTC) at which the file was
generated.

In addition there have been some minor changes to various other files to
support the new generation script.