Fix version comparisons for --api.
Previously, if generator.py was passed --api=1.3 it would generate a
table with only new commands found in Vulkan 1.3 and not Vulkan 1.2 or
lower. Obviously, this isn't very useful. This changes version
comparisons so that --api=X.Y will generate a table for all commands
found in Vulkan X.Y or lower. I'm not sure that this will continue to
make sense for some hypothetical Vulkan 2.0 but it's correct for Vulkan
<=1.3.