Change specification search behavior.
Previously, if generator.py wasn't provided with a specific XML spec
then it would download the most recent spec from Github. I don't think
this is especially safe because xml.etree is vulnerable to certain XML
expansion attacks. On top of that, it assumes you want the latest spec
and that Github is simply infalible.

Instead of downloading a file, the new implmentation will search a set
of predefined locations and default to the current working directory if
no spec file is found.

If, for whatever reason, the latest spec is needed then it's trivial to
grab yourself anyway.