Fix MVK replaced with MVKFL.
Previously, the header templates would replace any instance of VK in an
extension name with VKFL. This is apparently for every recognized
extension except 3 MoltenVK extensions. In the case of
VK_MVK_ios_surface, VK_MVK_macos_surface, and VK_MVK_moltenvk the MVK
was being inappropriately replaced with MVKFL. This is now fixed. Only
the first instance of VK found in an extension is now replaced.